refactor: Crew에서 수행하던 findByDate의 유효성 검사 메서드 Attendances에서 수행하도록 변경
